moved table to local test cases, using reflect.DeepEqual now
